added derivation of blueheart theme
git-svn-id: svn://code.dyne.org/veejay/trunk@887 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-1.0/veejay-themes/INSTALL.sh
@@ -24,11 +24,26 @@ mkdir -p $lndir 2>/dev/null
 for rcfile in $srcdir/*.rc; do
 	tmp=`basename $rcfile`
 	themename=`echo $tmp|cut -d '.' -f1`
-	mkdir $themedir/$themename 2>/dev/null
-	mkdir $lndir/$themename 2> /dev/null
-	if cp $rcfile $themedir/$themename/; then
-		ln -s $themedir/$themename/$tmp $lndir/$themename/gveejay.rc 2>/dev/null
-		echo "Installed $themename to $themedir"
+	extra=$themename.tar.bz2
+	mkdir -p $themedir/$themename 2>/dev/null
+	if test -f $extra ; then
+		cd $themedir/$themename/
+		if tar -jxvf $srcdir/$extra >/dev/null ; then
+			echo "   Extracted $extra in `pwd`"
+		fi	
+		echo "$themedir/$themename"
+		if cp $rcfile $themedir/$themename/; then
+			ln -s $themedir/$themename $lndir/$themename 2>/dev/null
+			ln -s $lndir/$themename/$tmp $themedir/$themename/gveejay.rc 2>/dev/null
+		fi
+		cd $srcdir		
+	else
+		mkdir $lndir/$themename 2>/dev/null
+		if cp $rcfile $themedir/$themename/; then
+			ln -s $themedir/$themename/$tmp $lndir/$themename/gveejay.rc 2>/dev/null
+			echo "Installed $themename to $themedir"
+		fi
+
 	fi
 done
